all crud video operations

+++ TikTokFinalProject/src/main/java/com/tiktok/model/repository/SoundRepository.java
@@ -1,10 +1,11 @@
 package com.tiktok.model.repository;
 
+import com.tiktok.model.entities.Sound;
+import com.tiktok.model.entities.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SoundRepository extends JpaRepository <Sound, Long> {
-
+public interface SoundRepository extends JpaRepository <Sound, Integer> {
 
 }
